data_ingest/requirements.txt
langchain
langchain-community
langchain-openai
qdrant-client
tiktoken
openai
python-dotenv
unstructured
unstructured[md]
sentence-transformers
